+++ src/unfold/sites.py
@@ -12,21 +12,12 @@ from django.core.paginator import Paginator
 from django.core.validators import EMPTY_VALUES
 from django.http import HttpRequest, HttpResponse
 from django.template.response import TemplateResponse
-from django.urls import URLPattern, path, reverse, reverse_lazy
+from django.urls import URLPattern, URLResolver, path, reverse, reverse_lazy
 from django.utils.encoding import force_bytes
 from django.utils.functional import lazy
 from django.utils.module_loading import import_string
 
 from unfold.dataclasses import DropdownItem, Favicon, SearchResult
-
-try:
-    from django.contrib.auth.decorators import login_not_required
-except ImportError:
-
-    def login_not_required(func: Callable) -> Callable:
-        return func
-
-
 from unfold.settings import get_config
 from unfold.utils import convert_color
 
@@ -47,7 +38,7 @@ class UnfoldAdminSite(AdminSite):
         elif self.login_form is None:
             self.login_form = AuthenticationForm
 
-    def get_urls(self) -> list[URLPattern]:
+    def get_urls(self) -> list[URLResolver | URLPattern]:
         extra_urls = []
 
         if hasattr(self, "extra_urls") and callable(self.extra_urls):
@@ -249,7 +240,7 @@ class UnfoldAdminSite(AdminSite):
 
     def search(
         self, request: HttpRequest, extra_context: dict[str, Any] | None = None
-    ) -> TemplateResponse:
+    ) -> TemplateResponse | HttpResponse:
         start_time = time.time()
 
         CACHE_TIMEOUT = 5 * 60
@@ -316,6 +307,7 @@ class UnfoldAdminSite(AdminSite):
             request,
             template=template_name,
             context={
+                "search_term": search_term,
                 "page_obj": paginator,
                 "results": paginator.page(request.GET.get("page", 1)),
                 "page_counter": (int(request.GET.get("page", 1)) - 1) * PER_PAGE,
@@ -329,7 +321,7 @@ class UnfoldAdminSite(AdminSite):
 
     def password_change(
         self, request: HttpRequest, extra_context: dict[str, Any] | None = None
-    ) -> HttpResponse:
+    ) -> TemplateResponse:
         from django.contrib.auth.views import PasswordChangeView
 
         from unfold.forms import AdminOwnPasswordChangeForm
@@ -416,12 +408,15 @@ class UnfoldAdminSite(AdminSite):
         return request.user.has_perm(f"{app_label}.view_{model_name}")
     
     def _get_navigation_items(
-        self, request: HttpRequest, items: list[dict], tabs: list[dict] = None
+        self, request: HttpRequest, items: list[dict], tabs: list[dict] | None = None
     ) -> list:
         allowed_items = []
 
         for item in items:
             link = item.get("link")
+
+            if not link:
+                continue
 
             if "active" in item:
                 item["active"] = self._get_value(item["active"], request)
@@ -562,14 +557,14 @@ class UnfoldAdminSite(AdminSite):
         return False
 
     def _get_is_tab_active(
-        self, request: HttpRequest, tabs: list[dict], link: str
+        self, request: HttpRequest, tabs: list[dict], nav_link: str
     ) -> bool:
-        for tab in tabs:
+        for tab_opts in tabs:
             has_primary_link = False
             has_tab_link_active = False
 
-            for tab_item in tab["items"]:
-                if link == tab_item["link"]:
+            for tab_item in tab_opts["items"]:
+                if nav_link == tab_item["link"]:
                     has_primary_link = True
                     continue
 
@@ -584,7 +579,7 @@ class UnfoldAdminSite(AdminSite):
 
         return False
 
-    def _get_config(self, key: str, *args) -> Any:
+    def _get_config(self, key: str, *args: Any) -> Any:
         config = get_config(self.settings_name)
 
         if key in config and config[key]:
@@ -604,7 +599,7 @@ class UnfoldAdminSite(AdminSite):
 
         return images
 
-    def _get_colors(self, key: str, *args) -> dict[str, dict[str, str]]:
+    def _get_colors(self, key: str, *args: Any) -> dict[str, dict[str, str]]:
         colors = self._get_config(key, *args)
 
         for name, weights in colors.items():
@@ -616,7 +611,7 @@ class UnfoldAdminSite(AdminSite):
 
         return colors
 
-    def _get_list(self, key: str, *args) -> list[Any]:
+    def _get_list(self, key: str, *args: Any) -> list[Any]:
         items = get_config(self.settings_name)[key]
 
         if isinstance(items, list):
@@ -624,7 +619,7 @@ class UnfoldAdminSite(AdminSite):
 
         return []
 
-    def _get_favicons(self, key: str, *args) -> list[Favicon]:
+    def _get_favicons(self, key: str, *args: Any) -> list[Favicon]:
         favicons = self._get_config(key, *args)
 
         if not favicons:
@@ -640,7 +635,7 @@ class UnfoldAdminSite(AdminSite):
             for item in favicons
         ]
 
-    def _get_site_dropdown_items(self, key: str, *args) -> list[dict[str, Any]]:
+    def _get_site_dropdown_items(self, key: str, *args: Any) -> list[DropdownItem]:
         items = self._get_config(key, *args)
 
         if not items: