+++ src/unfold/widgets.py
@@ -1,5 +1,4 @@
 import json
-import warnings
 from collections.abc import Callable
 from typing import Any
 
@@ -137,22 +136,29 @@ PROSE_CLASSES = [
     "font-normal",
     "whitespace-normal",
     "prose-sm",
+    "prose-a:text-link",
     "prose-blockquote:border-l-4",
     "prose-blockquote:not-italic",
     "prose-pre:bg-base-50",
     "prose-pre:rounded-default",
-    "prose-headings:font-medium",
-    "prose-a:text-primary-600",
-    "prose-headings:font-medium",
-    "prose-headings:text-base-700",
+    "prose-headings:font-semibold",
+    "prose-headings:text-important",
+    "prose-headings:tracking-tight",
+    "prose-strong:font-semibold",
     "prose-ol:list-decimal",
+    "prose-ol:flex",
+    "prose-ol:flex-col",
+    "prose-ol:gap-0.5",
+    "prose-ol:mt-2",
     "prose-ul:list-disc",
-    "prose-strong:text-base-700",
+    "prose-ul:flex",
+    "prose-ul:flex-col",
+    "prose-ul:gap-0.5",
+    "prose-ul:mt-2",
+    "prose-li:m-0",
     "dark:prose-pre:bg-base-800",
     "dark:prose-blockquote:border-base-700",
     "dark:prose-blockquote:text-base-300",
-    "dark:prose-headings:text-base-200",
-    "dark:prose-strong:text-base-200",
 ]
 
 CHECKBOX_CLASSES = [
@@ -800,16 +806,6 @@ class UnfoldAdminCheckboxSelectMultipleWidget(CheckboxSelectMultiple):
         }
 
 
-class UnfoldAdminCheckboxSelectMultiple(UnfoldAdminCheckboxSelectMultipleWidget):
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        warnings.warn(
-            "UnfoldAdminCheckboxSelectMultiple is deprecated and will be removed in a future release. "
-            "Please use UnfoldAdminCheckboxSelectMultipleWidget instead.",
-            stacklevel=2,
-        )
-        super().__init__(*args, **kwargs)
-
-
 class UnfoldBooleanWidget(CheckboxInput):
     def __init__(
         self, attrs: dict[str, Any] | None = None, check_test: Callable | None = None
@@ -882,16 +878,6 @@ class UnfoldAdminPasswordWidget(PasswordInput):
 
 class UnfoldAdminPasswordToggleWidget(UnfoldAdminPasswordWidget):
     template_name = "unfold/widgets/password_toggle.html"
-
-
-class UnfoldAdminPasswordInput(UnfoldAdminPasswordWidget):
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        warnings.warn(
-            "UnfoldAdminPasswordInput is deprecated and will be removed in a future release. "
-            "Please use UnfoldAdminPasswordWidget instead.",
-            stacklevel=2,
-        )
-        super().__init__(*args, **kwargs)
 
 
 class AutocompleteWidgetMixin: